Version 1.0.1 of DSS is adapted to Apache of linkis #444
